rename files to use coin acronyms
