+++ packages/blockchain-wallet-v4-frontend/src/scenes/Settings/Preferences/BtcLinkHandling/template.js
