+++ packages/blockchain-wallet-v4/src/utils/btc.spec.js
@@ -1,4 +1,4 @@
-import * as utils from './bitcoin'
+import * as utils from './btc'
 
 const fromHex = hex => Buffer.from(hex, 'hex')
