+++ packages/blockchain-wallet-v4/src/utils/index.js
@@ -1,6 +1,6 @@
 import * as bch from './bch'
-import * as bitcoin from './bitcoin'
-import * as ethereum from './ethereum'
+import * as bitcoin from './btc'
+import * as ethereum from './eth'
 import * as checks from './checks'
 
 export { bch, bitcoin, ethereum, checks }
